move script info
